+++ sancho-src/src/sancho/view/search/result/ResultDetailDialog.java
@@ -2,7 +2,6 @@ package sancho.view.search.result;
 
 import java.util.Arrays;
 import org.eclipse.jface.dialogs.Dialog;
-import org.eclipse.jface.dialogs.IDialogConstants;
 import org.eclipse.swt.graphics.GC;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.widgets.Composite;
@@ -29,7 +28,7 @@ public class ResultDetailDialog extends Dialog {
    }
 
    protected void createButtonsForButtonBar(Composite var1) {
-      this.createButton(var1, 0, IDialogConstants.OK_LABEL, true);
+      this.createButton(var1, 0, SResources.getString("b.ok"), true);
    }
 
    protected Control createDialogArea(Composite var1) {