+++ sancho-src/src/sancho/view/utility/dialogs/BandwidthDialog$ConfigOptionsDialog.java
@@ -2,7 +2,6 @@ package sancho.view.utility.dialogs;
 
 import java.util.Arrays;
 import org.eclipse.jface.dialogs.Dialog;
-import org.eclipse.jface.dialogs.IDialogConstants;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
@@ -31,8 +30,8 @@ class BandwidthDialog$ConfigOptionsDialog extends Dialog {
    }
 
    protected void createButtonsForButtonBar(Composite var1) {
-      this.createButton(var1, 0, IDialogConstants.OK_LABEL, true);
-      this.createButton(var1, 1, IDialogConstants.CANCEL_LABEL, false);
+      this.createButton(var1, 0, SResources.getString("b.ok"), true);
+      this.createButton(var1, 1, SResources.getString("b.cancel"), false);
    }
 
    protected Control createDialogArea(Composite var1) {